requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
google-generativeai>=0.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
